[HUDI-6644] Flink append mode use auto key generator, refactor
